move utils file to its own folder
